+++ backend/benefit/applications/tests/test_applications_api.py
@@ -566,7 +566,7 @@ def test_application_post_invalid_data(
     assert len(Application.objects.all()) == 0
 
     del data["id"]
-    data["de_minimis_aid_set"][0]["amount"] = "300000.00"  # value too high
+    data["de_minimis_aid_set"][0]["amount"] = "300001.00"  # value too high
     data["status"] = "foo"  # invalid value
     data["bases"] = ["something_completely_different"]  # invalid value
     data["applicant_language"] = None  # non-null required
@@ -738,7 +738,7 @@ def test_application_put_read_only_fields(api_client, application):
 
 def test_application_put_invalid_data(api_client, application):
     data = ApplicantApplicationSerializer(application).data
-    data["de_minimis_aid_set"][0]["amount"] = "300000.00"  # value too high
+    data["de_minimis_aid_set"][0]["amount"] = "300001.00"  # value too high
     data[
         "status"
     ] = ApplicationStatus.ACCEPTED  # invalid value when transitioning from draft
@@ -822,7 +822,7 @@ def test_application_edit_de_minimis_aid_too_high(api_client, application):
 
     previous_aid = copy.deepcopy(data["de_minimis_aid_set"])
     data["de_minimis_aid"] = True
-    data["de_minimis_aid_set"][0]["amount"] = "150000"
+    data["de_minimis_aid_set"][0]["amount"] = "250000"
     data["de_minimis_aid_set"][1]["amount"] = "50001"
 
     response = api_client.put(
@@ -1353,7 +1353,7 @@ def test_application_status_change_as_handler(
         if to_status == ApplicationStatus.ADDITIONAL_INFORMATION_NEEDED:
             assert application.messages.count() == 1
             assert (
-                "Please make the corrections by 18.06.2021"
+                "Please make the corrections by 11.06.2021"
                 in application.messages.first().content
             )
             assert len(mailoutbox) == 1
@@ -2167,7 +2167,7 @@ def test_application_additional_information_needed_by(api_client, handling_appli
         )
     response = api_client.get(get_detail_url(handling_application))
     assert response.status_code == 200
-    assert response.data["additional_information_needed_by"] == date(2021, 12, 15)
+    assert response.data["additional_information_needed_by"] == date(2021, 12, 8)
 
 
 def test_application_status_last_changed_at(api_client, handling_application):