+++ backend/kesaseteli/applications/tests/test_youth_applications_api.py
@@ -16,7 +16,7 @@ from django.contrib.auth import REDIRECT_FIELD_NAME
 from django.contrib.auth.models import AnonymousUser
 from django.core import mail
 from django.test import override_settings
-from django.utils import timezone
+from django.utils import timezone, translation
 from django.utils.html import strip_tags
 from django.utils.timezone import localdate
 from freezegun import freeze_time
@@ -65,8 +65,33 @@ from shared.common.tests.conftest import (
     superuser_client,
 )
 from shared.common.tests.factories import UserFactory
+from shared.common.tests.names import INVALID_NAMES, VALID_NAMES
 from shared.common.tests.test_validators import get_invalid_postcode_values
 from shared.common.tests.utils import normalize_whitespace
+
+# YouthApplication's fields that are validated as names
+YOUTH_APPLICATION_NAME_FIELDS = ["first_name", "last_name", "school"]
+
+# Mandatory fields of YouthSummerVoucher in youth summer voucher email
+MANDATORY_YOUTH_SUMMER_VOUCHER_FIELDS_IN_VOUCHER_EMAIL = [
+    "employer_summer_voucher_application_end_date_localized_string",
+    "min_work_compensation_with_euro_sign",
+    "min_work_hours",
+    "summer_job_period_localized_string",
+    "summer_voucher_serial_number",
+    "voucher_value_with_euro_sign",
+    "year",
+]
+
+# Mandatory fields of YouthApplication in youth summer voucher email
+MANDATORY_YOUTH_APPLICATION_FIELDS_IN_VOUCHER_EMAIL = [
+    "email",
+    "first_name",
+    "last_name",
+    "phone_number",
+    "postcode",
+    "school",
+]
 
 
 def create_same_person_previous_year_accepted_application(
@@ -1290,6 +1315,57 @@ def test_youth_application_post_valid_random_data(  # noqa: C901
     NEXT_PUBLIC_DISABLE_VTJ=True,
 )
 @pytest.mark.django_db
+@pytest.mark.parametrize("name_field", YOUTH_APPLICATION_NAME_FIELDS)
+@pytest.mark.parametrize("name", VALID_NAMES)
+def test_youth_application_post_valid_non_ascii_names(api_client, name_field, name):
+    youth_application = YouthApplicationFactory.build()
+    data = YouthApplicationSerializer(youth_application).data
+    data[name_field] = name
+    response = api_client.post(get_list_url(), data)
+
+    assert response.status_code == status.HTTP_201_CREATED
+    assert "id" in response.data
+    created_youth_application = YouthApplication.objects.get(pk=response.data["id"])
+    assert getattr(created_youth_application, name_field) == name
+
+
+@override_settings(
+    NEXT_PUBLIC_MOCK_FLAG=False,
+    NEXT_PUBLIC_DISABLE_VTJ=True,
+)
+@pytest.mark.django_db
+@pytest.mark.parametrize("name_field", YOUTH_APPLICATION_NAME_FIELDS)
+@pytest.mark.parametrize("name", VALID_NAMES)
+def test_youth_application_post_names_with_whitespace(api_client, name_field, name):
+    youth_application = YouthApplicationFactory.build()
+    data = YouthApplicationSerializer(youth_application).data
+    whitespace = "\t\r\n  "
+    data[name_field] = whitespace + name + whitespace
+    response = api_client.post(get_list_url(), data)
+
+    assert response.status_code == status.HTTP_201_CREATED
+    assert "id" in response.data
+    created_youth_application = YouthApplication.objects.get(pk=response.data["id"])
+    assert getattr(created_youth_application, name_field) == name.strip()
+
+
+@pytest.mark.django_db
+@pytest.mark.parametrize("name_field", YOUTH_APPLICATION_NAME_FIELDS)
+@pytest.mark.parametrize("name", INVALID_NAMES)
+def test_youth_application_post_invalid_names(api_client, name_field, name):
+    youth_application = YouthApplicationFactory.build()
+    data = YouthApplicationSerializer(youth_application).data
+    data[name_field] = name
+    response = api_client.post(get_list_url(), data)
+
+    assert response.status_code == status.HTTP_400_BAD_REQUEST
+
+
+@override_settings(
+    NEXT_PUBLIC_MOCK_FLAG=False,
+    NEXT_PUBLIC_DISABLE_VTJ=True,
+)
+@pytest.mark.django_db
 def test_youth_application_post_valid_audit_log(api_client):
     youth_application = YouthApplicationFactory.build()
     data = YouthApplicationSerializer(youth_application).data
@@ -1676,6 +1752,60 @@ def test_youth_summer_voucher_email_plaintext_html_similarity(api_client, langua
         ).ratio()
         >= 0.9
     ), "Email's plaintext and HTML content should be very similar"
+
+
+@pytest.mark.django_db
+@override_settings(
+    NEXT_PUBLIC_MOCK_FLAG=True,
+    EMAIL_BACKEND="django.core.mail.backends.locmem.EmailBackend",
+)
+@pytest.mark.parametrize("language", get_supported_languages())
+def test_youth_summer_voucher_email_html_content(api_client, language):
+    acceptable_youth_application = AcceptableYouthApplicationFactory(language=language)
+    api_client.patch(
+        get_accept_url(acceptable_youth_application.pk),
+        data=json.dumps(get_test_handling_data()),
+        content_type="application/json",
+    )
+    assert len(mail.outbox) > 0
+    youth_summer_voucher_email = mail.outbox[-1]
+    html_msg = youth_summer_voucher_email.alternatives[0][0]
+    acceptable_youth_application.refresh_from_db()
+    voucher = acceptable_youth_application.youth_summer_voucher
+
+    with translation.override(acceptable_youth_application.language):
+        for field in MANDATORY_YOUTH_SUMMER_VOUCHER_FIELDS_IN_VOUCHER_EMAIL:
+            assert str(getattr(voucher, field)) in html_msg
+
+        for field in MANDATORY_YOUTH_APPLICATION_FIELDS_IN_VOUCHER_EMAIL:
+            assert str(getattr(acceptable_youth_application, field)) in html_msg
+
+
+@pytest.mark.django_db
+@override_settings(
+    NEXT_PUBLIC_MOCK_FLAG=True,
+    EMAIL_BACKEND="django.core.mail.backends.locmem.EmailBackend",
+)
+@pytest.mark.parametrize("language", get_supported_languages())
+def test_youth_summer_voucher_email_plaintext_content(api_client, language):
+    acceptable_youth_application = AcceptableYouthApplicationFactory(language=language)
+    api_client.patch(
+        get_accept_url(acceptable_youth_application.pk),
+        data=json.dumps(get_test_handling_data()),
+        content_type="application/json",
+    )
+    assert len(mail.outbox) > 0
+    youth_summer_voucher_email = mail.outbox[-1]
+    plaintext_msg = youth_summer_voucher_email.body
+    acceptable_youth_application.refresh_from_db()
+    voucher = acceptable_youth_application.youth_summer_voucher
+
+    with translation.override(acceptable_youth_application.language):
+        for field in MANDATORY_YOUTH_SUMMER_VOUCHER_FIELDS_IN_VOUCHER_EMAIL:
+            assert str(getattr(voucher, field)) in plaintext_msg
+
+        for field in MANDATORY_YOUTH_APPLICATION_FIELDS_IN_VOUCHER_EMAIL:
+            assert str(getattr(acceptable_youth_application, field)) in plaintext_msg
 
 
 @pytest.mark.django_db
@@ -2453,6 +2583,7 @@ def test_youth_applications_set_excess_additional_info(
                 [AdditionalInfoUserReason.OTHER],
                 " \tLeading\n and trailing whitespace should get removed \n\t ",
             ),
+            *[([AdditionalInfoUserReason.OTHER], name) for name in VALID_NAMES],
         ]
     ],
 )