+++ frontend/benefit/applicant/src/components/applications/applicationList/listItem/ListContents.tsx
@@ -1,0 +1,91 @@
+import {
+  $Heading,
+  $HeadingContainer,
+  $ListWrapper,
+  $OrderByContainer,
+} from 'benefit/applicant/components/applications/applicationList/ApplicationList.sc';
+import ListItem from 'benefit/applicant/components/applications/applicationList/listItem/ListItem';
+import { ApplicationListProps } from 'benefit/applicant/components/applications/applicationList/useApplicationList';
+import { Select } from 'hds-react';
+import React, { PropsWithChildren, useEffect } from 'react';
+import LoadingSkeleton from 'react-loading-skeleton';
+import Container from 'shared/components/container/Container';
+import theme from 'shared/styles/theme';
+import { OptionType } from 'shared/types/common';
+
+interface ListContentProps {
+  headingText: React.ReactNode;
+  orderByOptions?: OptionType[];
+  status: string[];
+  items: JSX.Element[];
+  noItemsText?: React.ReactNode;
+  onListLengthChanged?: (isLoading: boolean, length: number) => void;
+}
+
+type Props = PropsWithChildren<
+  Omit<ApplicationListProps, 'language'> & ListContentProps
+>;
+
+const ListContents = ({
+  shouldShowSkeleton,
+  shouldHideList,
+  headingText,
+  orderByOptions,
+  orderBy,
+  setOrderBy,
+  t,
+  status,
+  noItemsText,
+  items,
+  hasItems,
+  list,
+  onListLengthChanged,
+  children,
+}: Props): JSX.Element => {
+  useEffect(() => {
+    onListLengthChanged?.(shouldShowSkeleton, list.length);
+  }, [shouldShowSkeleton, list.length, onListLengthChanged]);
+
+  if (shouldHideList && !noItemsText) return null;
+
+  return (
+    <Container backgroundColor={theme.colors.silverLight}>
+      {shouldShowSkeleton && (
+        <>
+          <$HeadingContainer>
+            <LoadingSkeleton width="20%" />
+          </$HeadingContainer>
+          <$ListWrapper>
+            <ListItem isLoading />
+          </$ListWrapper>
+        </>
+      )}
+      {!shouldShowSkeleton && (
+        <>
+          <$HeadingContainer>
+            <$Heading>{headingText}</$Heading>
+            <$OrderByContainer>
+              {orderByOptions?.length > 1 && (
+                <Select<OptionType>
+                  id={`application-list-'${status.join('-')}-order-by`}
+                  options={orderByOptions}
+                  defaultValue={orderBy}
+                  onChange={setOrderBy}
+                  label={t('common:applications.list.common.sortOrder')}
+                  disabled={!hasItems}
+                />
+              )}
+            </$OrderByContainer>
+          </$HeadingContainer>
+          <$ListWrapper>
+            {items}
+            {!hasItems && noItemsText}
+          </$ListWrapper>
+          {children}
+        </>
+      )}
+    </Container>
+  );
+};
+
+export default ListContents;