+++ frontend/benefit/applicant/src/constants.ts
@@ -18,7 +18,7 @@ export enum ROUTES {
   DECISIONS = '/decisions',
 }
 
-export const MAX_DEMINIMIS_AID_TOTAL_AMOUNT = 200_000;
+export const MAX_DEMINIMIS_AID_TOTAL_AMOUNT = 300_000;
 
 export enum SUPPORTED_LANGUAGES {
   FI = 'fi',