+++ frontend/benefit/shared/src/constants.ts
@@ -108,7 +108,7 @@ export enum DE_MINIMIS_AID_KEYS {
 
 export enum DE_MINIMIS_AID {
   // eslint-disable-next-line unicorn/numeric-separators-style
-  MAX_AMOUNT = 200000,
+  MAX_AMOUNT = 300000,
 }
 
 export enum CALCULATION_EMPLOYMENT_KEYS {